Introduce the HardwareManager CRD
The HardwareManager CRD defines an instance of a hardware manager, which
is used by the hwmgr-plugin when processing a NodePool request to
determine which adaptor to use, as well as provide adaptor-specific
configuration data to the adaptor.

Each adaptor defines its own reconciler for watching HardwareManager
CRs, filtered by its own adaptor ID. This ensures each reconciler is
only kicked to process its own CRs.

The NodePool reconciler gets the corresponding HardwareManager CR, based
on the .spec.hwMgrId field, and passes both the NodePool and
HardwareManager CRs to the appropriate adaptor.

Not yet implemented:
- HardwareManager reconcilers will be responsible for setting
  appropriate Conditions in the CR status when validating the CR.
- Add real configuration fields to the adaptor-specific data structures,
  and establish a connection to the hardware manager (if appropriate)/

Signed-off-by: Don Penney <dpenney@redhat.com>