Fix missing code change in the initial submission
This commit restores a missing code change that was inadvertently
lost in the initial metal3 plugin submission.

The issue was likely caused by reverting some untested code before
pushing the previous PR, which caused it to be absent in the test
environment but appear in the main branch after the PR merge.

Signed-off-by: Tao Liu <tali@redhat.com>